[GH] Add collection playground project
Note that I decided to not include collections2 since it is just an
infra project and we've decided not to merge it into collection as is
(instead do it piecemeal).

Bug: n/a
Test: gw bOS

This is an imported pull request from https://github.com/androidx/androidx/pull/330.

Resolves #330
Github-Pr-Head-Sha: 01e2bddd6be4d6f4ca60977a6277ea93b8dea391
GitOrigin-RevId: 285741a43aba4a0166c2cf7ba9613e8a107ada52
Change-Id: I9e1b0f84f8572577af33bff2eaf2bfbe9c1ae305