Update Quirks for incorrect MediaCodecInfo issue
The Redmi note 4 & LG K430 reports the maximum supported video encoder
resolution is 1920x1072, however, we can successfully open the encoder
with 1920x1080. So we can treat the devices that report the incorrect
MediaCodecInfo.

Remove CamcorderProfileResolutionNotSupportedByEncoderQuirk,
and add the above devices into the MediaCodecInfoReportIncorrectInfoQuirk.
It will ignore VideoCapability check and force open the avc video
encoder with FHD resolution.

Bug: 216583006
Test: SupportedQualitiesVerificationTest
Change-Id: I5e6b293969a44e379ea37e1dfe2bb32e73ccfa75