Move ExifOnly check to where the data is read
The previous check caused all subsequent data to be read via a
BufferedInputStream with a buffer size of 6 bytes, which seemed
undesirable.

isExifDataOnly was also assuming BufferedInputStream.read will
read data in full, which isn't technically guaranteed.
BufferedInputStream will make a best attempt to read the full
amount, but will not block to do so.

Finally, getStandaloneAttributes was assuming that the input
stream would indicate all remaining data was available, which is
also not guaranteed.

Test: ExifInterfaceTest
Bug: 220181775
Change-Id: Id7b73d9f49b459d44b0794cabdb94654e475a4f8