SplitToggleChip UX Spec Changes
The UX spec for the SplitToggleChip has been updated and so a new ToggleChip.splitChipColors method has been added to give developers colors to match the updated specification and easily customize the appropriate fields.

Bug: 221000735
RelNote: "The recommended coloring for the SplitToggleChip has been updated. SplitToggleChip's now have a solid colored background when checked or unchecked with the coloring of the ToggleControl being the primary indication of whether the component is checked or not. We have added a new ToggleDefaults.splitToggleChipColors() to support the new color scheme. We have also simplified the toggleChipColors() methods removing the splitBackgroundOverlayColor"
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I7e66e3c9a7c74b588edd7db9bca22f807aca68fc