Fix flaky test testSetAccessibilityPaneTitle_sendsOutCorrectEvent
There are two errors in the test: 1. event.getSource is not cached
in the test, so the nullness check may not always work. 2. We
call createAccessibilityNodeInfo() in a non-ui thread, and this
caused race conditions.

Instead of using source comparision, use pane title comparision
instead, as other tests do.

Also changed the setAccessibilityPaneTitle() and setStateDescription
to ui thread.

Bug: 206644987

Test: tests pass locally. Will check post submit to determine
whether the test is still flaky.

Change-Id: I0cd16f58a1a231c5d4b4a19a2d87d0f45ea712d4