Extend Picker to have a read-only mode
Added readOnly and label parameters to support a read-only mode for Picker. When readOnly = true, the currently selected option is displayed. There is a slot for a label as well - if that is supplied, the label is also displayed above the currently selected option in read-only mode. This new mode is intended to be used for screens that display multiple Pickers, only one of which is editable at a time, the others being shown as read-only.

Test: androidx.wear.compose.material.test

Relnote: "We have added a read-only mode to Picker, for screens with multiple Pickers where only one Picker is editable at a time. When the Picker is read-only, it displays the currently selected option and a label if one has been provided."

Bug: 222267781
Change-Id: I879de7842a21d000d5a2554922be976899f930d3