Update scaling params on ScalingLazyColumn as the UX Spec has been updated
The UX spec for the scaling of list items in the ScalingLazyColumn has been updated. Adjusting the parameters accordingly.

RelNote: "We have updated the default scaling params of the ScalingLazyColumn in-line with the latest Wear Material Design UX specifications. Visually this results in list items starting to be scaled closer to the list center, but being less scaled at the list edge than before. "
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 220842937
Change-Id: Ica8f35ee3860ef8ab48d71a52fb28383597b2ecb