Fixed FrameData equality problem, added tests and docs
Previously, the equals() function in the FrameData
subclasses assumed that the 'other' object was of the same
type, and auto-cast it, causing a crash if it wasn't of that
type.

New code checks whether the class is the same (returning false
for equals() if it is not). Also, added test for equality to
JankStatsTest to check for these failures as well as others
on equality and inequality.

Also, added/fixed docs for FrameData and its subclasses.

Bug: 218296544, 218806612
Test: JankStatsTest#testEquality()
Change-Id: I1a2df4d3fd03b052bc90a0a174719475f5d216c6