Refine startTime value
Two minor fixes to refine the time used to make the start time
of a frame:
- INTENDED_VSYNC_TIMESTAMP: On API 26+, this time more accurately
represents when the frame should have started than the previous
VSYNC_TIMESTAMP. They are usually the same, but in janky times,
the INTENDED time can capture more detail about when things should
have occurred, and how long they actually took.
- current vs previous frames: The use of INTENDED_VSYNC_TIMESTAMP
means that some frames can appear to overlap (because the system
tried to start one frame while the previous was still being processed).
This can lead to errors in state captures for those frames since they
appear in multiple frames for the same point in time. To avoid that
problem, the system now caches the previous "end" time of a frame,
and avoids using a start time less than that value.

Bug: 213245198
Test: Passed all JankStats tests on multiple API versions
Change-Id: I23fcc474f48619be9e0a996d1d0ac4d8e168e9e1