includeFontPadding line height fix
When we disabled includeFontPadding, because of the differences
between platform versions, we updated LineHeightSpan not to apply
line height to first line. This might be causing small rounding (1px)
differences between the line heights of 1st and the rest of the lines.

This CL updates AndroidParagraphHelper to return to the state before
includeFontPadding changes if includeFontPadding is true.

Finally, canvas.translate is not needed to be called if the
topPadding is 0. This CL updates TextLayout.kt not to translate if
topPadding is 0.

Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: Manual Test in demo app

Bug: 171394808
Bug: 227095468

Change-Id: I112732a542b47b373d6b45992b9758f1bed0b668