TypefaceCompatApi29 finds best from requested style
This causes fonts closer to the desired style to be chosen on
resolution.

Test: ./gradlew :core:core:cAT
Bug: b/194553426
Relnote: "TypefaceCompat will now respect both requested and loaded
style information on API 29+. This is a behavior change when the loaded
fonts are not FontWeight.Normal or FontWeight.Bold, as the actual loaded
weight and style will be used. For more information see b/194553426"

Change-Id: I1f90279afa68afa450bb98eb7c5237464c5f57d8