Add ripple effect to clickable UI elements
This change sets the foreground to "selectableItemBackground" for all
views except buttons so that they display a ripple when they are
clicked (only if an onClick action is set / the view is clickable). This
is not necessary for buttons since they have a ripple background by
default.

This matches the Compose UI behavior which sets a ripple background for
every element that has Modifier.clickable() set on it.

Bug: 213389535
Test: set Modifier.clickable() on an Image or Column, verify that there
 is a ripple on the view
Change-Id: Ia32535c0143c94ac97a55faee669119cbae47d5b