Add in/out transitions for Dialog
Added in/out transitions for Dialog. The visibility of the Dialog is now controlled by a showDialog parameter - the caller can set showDialog=false in an action (e.g. a Button click), which triggers the out transition and then hides the dialog. Note that the out animation is not triggered when leaving the Dialog using swipe-to-dismiss, because that has its own animations.

Also added a default value for the state parameter in the SwipeToDismissBox overload that accepts an onDismissed parameter (because typically, the state is no longer needed outside the SwipeToDismissBox, since the LaunchedEffect to dismiss is already handled).

Test: Run androidx.wear.compose.material.test
Bug: 220853895
Relnote: "We have implemented in/out transitions for Dialog. A showDialog parameter has been added and the Dialog now controls its own visibility (this enables Dialog to run the intro and outro animations when the Dialog is shown/hidden). Note that the outro animation is not performed when the user leaves the dialog via swipe-to-dismiss. We have also added a default value for state in the recently added SwipeToDismissBox overload."
Change-Id: I682a03ea9354ed112f2b5be15634b27c41c6f31a