CompactChip - Updates to new UX Spec
Updates to CompactChip to bring it inline with latest UX Spec.
1. Font should be caption1 not button
2. ContentPadding should be 0.dp top and bottom and 12.dp start/end
3. For an icon only compact chip center align the icon
4. When both a label and an icon are shown the icon should be 20x20, when only icon is shown the icon should be 24x24
5. Label text should be start aligned when there is also an icon and center aligned when there is no icon, just a label

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have updated CompactChip to bring it inline with the latest UX Specification. Padding has been reduced to horizontal = 12.dp and vertical = 0.dp. The font for label has been changed from button to caption1. The recommended icon sizes are 20x20 when both icon and label are present and 24x24 for an icon only compact chip. For the icon only use case we have also ensured that the icon is center aligned."
Bug: 226939898
Change-Id: Iea2bef2ab11f750f9fe459c240354f06214f7930