Draw order & Snackbar placement for BottomSheetScaffold
Previously, we were drawing the layout's body content over the top bar's elevation, but we now respect the draw order.

BottomSheetScaffold now also takes the sheet's state into account when placing Snackbars: in the collapsed state, Snackbars are placed above the sheet and FAB; in the expanded state, Snackbars are anchored to the sheet's bottom.

We are now also explicit about the case of sheetHeight == peekHeight, only having a collapsed anchor/state in this case. We are looking to explore different structures (i.e. a list) to represent multiple associated values for anchors.

Relnote: Fixed a bug where BottomSheetScaffold was drawing over the top app bar's shadow. BottomSheetScaffold now also takes the sheet's state into account when placing Snackbars: in the collapsed state, Snackbars are placed above the sheet and FAB; in the expanded state, Snackbars are anchored to the sheet's bottom.

Bug: b/187771422
Test: Added a few tests around snackbar positioning and expected anchors in case sheetHeight == peekHeight, removed bottomSheetScaffold_topAppBarIsDrawnOnTopOfContent and created a screenshot test as replacement
Change-Id: Ia80b5d1ebcd4c86431eab22a985ab2b1fa824843