Add table names of key and value columns in @MapInfo
This allow users with ambiguous columns and tables alias to make it clear which table.column should be used as key or value via @MapInfo.

Note that this CL specifically does not use the table information to resolve the column index on a duplicate column query, but might do so with a latter change.

Relnote: "Added APIs for providing key and value tables names for disambiguation in @MapInfo"
Test: QueryMethodProcessorTest
Change-Id: Icc4b50a029d33c49fbfe7265d10b6be1b15da9c3