Wait 3A when using torch as flash
Some devices always report the FLASH_REQUIRED in the aeState after
the torch is opened. It would be unable to detect the AE converged signal.
This change will wait for the 3A converge for the above case, but not
treat the AE FLASH_REQUIRED state as a converged signal. It will wait for
at most 2 seconds for the auto exposure update.

AeState.UNKNOWN means it doesn't have valid AE info. For this kind of device,
we tend to wait for a few more seconds for the auto exposure update.
So the UNKNOWN state should not be treated as the AE converge signal.

Bug: 228800282
Test: CamXCoreAppFlashTorchCheckQRCode
Change-Id: Ia49e8e7228a201d59a25535c8c334bee36cc0cb0