Fix several GlesRenderer bugs
There was a bug in GlesRenderer.backgroundThreadInitInternal where
it was only calling onBackgroundThreadGlContextCreated if
EGL14.eglCreateContext was called. Watchfaces typically do GL
init in onBackgroundThreadGlContextCreated so this could lead to
crashes in the headless instance upon installation when the
system tried to take a screenshot.

This patch also 'fixes' another rather more mysterious bug where
there was a visual glitch in the screenshot caused by verticalFlip.
The problem is this glitch doesn't seem to occur in unit tests,
however screenshots taken with the new version of verticalFlip
seem to be correct...

Bug: 216722595
Test: Manual testing with partner watchface confirms absence of visual glitch in screenshot after patch applied
Change-Id: Ice0106ca17a793fc15784eeeb634a4f83291400a