Find a node by anchor hash to access parameters
This is a temporary hack to enable lazy parameter fetching by the
compose inspector agent.

We will eventually get a way to retrieve the `CompositionGroup`
from a given anchor. Until then find the parameters by:
Start building the tree again and save the node which has the
wanted anchor.

In this change: just add the capability for find a node by the
anchor hash, do not start using this feature.

Bug: 229657989
Test: Added unit test
Change-Id: I4aec197f9a35c544a211089d44f4c1bb2140a0d1