Added lint check for empty navDeepLink in Kotlin DSL.
Adding in a lint check to provide a warning to users that attempt to create a NavDeepLink using the Kotlin DSL navDeepLink function. Using an empty NavDeepLink will result in a runtime exception as a deep link with a url, action, and/or mimetype is invalid.

Relnote: Attempting to create an empty  NavDeepLink  using the  navDeepLink  Kotlin DSL will not result in a lint warning indicating that a deep link needs an uri, action, and/or mimetype to be valid.
Test: Included tests of API version and the lint rule
Bug: 154038883
Change-Id: I08d2feba9b90ee934b8834c5573914fa44d7dc45