Revert^3 "Adding implementation for Uwb ranging"
eaef4e45e23663a9d83164d9f50416d2809cafe6

Revert reason: checked in prebuilts are overriding a version that
is a version that already exists and it does not match the existing
maven artifact content of `com.google.android.gms:play-services-nearby:18.1.0`.
It is not clear if this is a locally built or a build server version.

Change-Id: Ic6691da3bf7a09e2af8e7db5982e14f9d11f7b37