Alleviate issue on PickerState.scrollToOption
When using PickerState.scrollToOption on a Picker with repeatedItems =
true that it's on readonly mode, it may fail to snap properly to an
item that wasn't draw the last time the Picker was visible (because
of an undelying issue on the ScalingLazyColumn).
To alleviate this, of the multiple posible items that correspond to
the option we want to selectwe pick the items closest to the current one
that has more chances of being on the screen on last render.
Relnote: N/A
Test: Added
Bug: 230485107

Change-Id: I600780826e502251f1d88450b2d7d657db9b0d91