Add a minimal watch face using multiple instances.
This commit adds a watch face that supports multiple instances.

This watch face uses the instance ID to store some local state that is
not part of the watch face user style.

This should only be done for watch faces that need configuration that
cannot be handled using the watch face user style, as this
configuration options will not be available on the phone integrated
configuration.

Bug: 231269324
Test: manual
Change-Id: I80f17ca51482445ecabad451fdee7955b3f192ee