Catch OutputConfiguration exceptions
When a CameraGraph is started, we create OutputConfigurations with the
given parameters, including the output Surfaces set. The problem is the
Surface can actually be destroyed by the caller while we're doing so,
causing crashes while creating an OutputConfiguration, and in turn
crashing the entire app.

We now attempt to handle this by catching the exceptions thrown by
OutputConfiguration. When that happens, we ignore the
OutputConfigurations that come with invalid Surfaces.

Bug: 230795097
Test: Open CameraX test app with CameraPipe backend and toggle Photo, \
  Analysis option multiple times.
Change-Id: I64102ff934aed0b9ea948ceca0b4c033e018cf69