Add key injection API for Compose UI tests
Adds a `KeyInjectionScope` similar to the existing Touch, Mouse and
Rotary injection scopes. Adds support for keyboard input injection
in `MultiModalInjectionScope` and `InputDispatcher` as well as a
platform-specific implementation of `KeyEvent` injection in
`AndroidInputDispatcher`.

The API in `KeyInjectionScope` is currently experimental and has
been annotated with `@ExperimentalTestApi` since it makes use of the
experimental `androidx.compose.ui.input.key` API for representing
keyboard keys in a platform independent way.

Bug: 229831515
Test: ./gradlew compose:ui:ui-test:test
      && ./gradlew compose:ui:ui-test:cC
Relnote: "Added experimental support for injecting key events. Use
`performKeyInput` to send key events, or send them through the `key`
property of `MultiModalInjectionScope` during a multi modal input
gesture with `performMultiModalInput`. See `KeyInjectionScope` for
documentation of the API."

Change-Id: Ic5000922484274309323ee4d13c9941e0d737673