Fix deadlock in CameraX ImageAnalysis
The root cause is there are one main thread to analyze image in app and one background thread to promote a cached image to analyze in CacheAnalyzingImageProxy, they are trying to wait for the other's acquired lock while holding the lock the other is waiting.

The analyzeCachedImage in background thread has acquired ImageAnalysisNonBlockingAnalyer#mLock and tried to acquire ImageAnalysisAbstractAnalyzer#mAnalyzerLock. Meanwhile the setAnalyzer in main thread has acquired ImageAnalysisAbstractAnalyzer#mAnalyzerLock and tried to acquire ImageAnalysisNonBlockingAnalyer#mLock in clearCache.

The solution is to break this circular waiting logic by moving clearCache out of ImageAnalysisAbstractAnalyzer#mAnalyzerLock's  critical section.n

Bug: b/216791440
Test: Run integration test app w/ and w/o this change to verify the deadlock issue cannot reproduce
Change-Id: Id78560872e25be2c165c5ef4381a195f6ef3b974