Add IntervalList api
Reshaped the api of previously internal IntervalList and MutableIntervalList so it is more clean and does not expose the internal binary search implementation on the users. Exposed it as a public experimental api so the external implementations of LazyLayout can use it.

Relnote: Added new experimental IntervalList and MutableIntervalList. It allows to represent some list of values via multiple intervals. It will be useful when you want to define your own dsl similar to the one used by LazyColumn where list items can be defined via multiple item/items calls.
Fixes: 228580728
Test: more tests in MutableIntervalListTest
Change-Id: I2d05e91646068eab522514024b483d88d77be309