Improve WindowInsets.ime kdoc to include API 23.
The WindowInsets.ime value is actually supported on API 23+,
so I updated the kdoc to say that. However, it's only *animated* on 30+,
so I clarified that in the kdoc as well.

This change also adds a demo to the demo app that shows the actual
insets.

Bug: b/230756508
Test: n/a, only a docs change
Relnote: "Clarified the documentation for `WindowInsets.ime` to state
    that `ime` insets are reported as far back as API 23, but only
    _animated_ on 30+."
Change-Id: Ia7fc002bde64074be7a176121483bff3017f24a8