Fix EGLExt/Spec/GLRenderer Docs
Updated docs to indicate failure cases
while attempting to create an EGLImageKHR
object from a HardwareBuffer instance.

Updated docs to indicate that consumers
of EGLImageKHR/EGLSyncKHR objects are to
release these objects by calling their
corresponding destructive APIs.

Fixed docs that referred to a Surface object
in createRenderTarget that is to be used
explicitly in scenarios where a Surface is not
necessary.

Updated EglManagerTest to query previously
missing EGLExtension to support obtaining
an EGLImageKHR object from a HardwareBuffer
instance.

Fixes: 233948725
Fixes: 233948878
Test: updated EglManagerTest
Change-Id: I1c87dca98a9932a11d75e26042a4184ae23d9d44