Have FragmentStateManager saveState return Bundle
Rather than FragmentStateManager directly accessing
the FragmentStore when we save state, we want to
simply return the state Bundle and then whenever
we call this function, if necessary, we will add
that state to the FragmentStore. This is done so
that we can access the state bundle without needing
access to the FragmentStore.

Test: ran all tests in androidx.fragment.app
Bug: 207158202
Change-Id: Ia335be8a46e376d4a622ebee972f897a596a51f6