Add a new KMP StudioType
This CL adds a new StudioType = KMP to include all KMP enabled project
(exclusing compose). We also turn on native build targets when it
is set.

This StudioType also enables native targets. It is rather temporary until
we have proper integration but the copy-pasta in build.gradle files was
growing big so I decided to add this as a temporary solution until we have
first class support in AndroidX infra for KMP projects.

Bug: n/a
Test: ANDROIDX_PROJECTS=kmp ./gradlew collection:collection:allTests
Change-Id: Ife91389cabb951fe32060d5e7ec126de1e0dc344