Reduce max popup/dialog elevation
See http://go/elevation-mitigation for more details.

Older Android versions are affected by a surface insets matrix bug
fixed in ag/11788480.  This causes the region passed into
AccessibilityInteractionController.adjustIsVisibleToUserIfNeeded to be
incorrectly offset, and the magnitude of the error scales with the
amount of window-level elevation.

Since maxSupportedElevation works by setting window-level elevation to
the max, then hiding the associated visual effect (so that compose-
rendered children can render their own elevation), a high max itself is
enough to trigger this bug, even if the actual visible elevation is much
lower.  With 30.dp of max elevation, the region is so offset it usually
stops intersecting with the bottom element of a dropdown popup entirely,
causing the element to be hidden from TalkBack.

At a maximum of 8.dp, this system bug is mitigated to the same level as
the View system, and would only affect only tiny elements squeezed near
the bottom of custom Dialogs.

Bug: 232788477
Bug: 236109671
Test: Android Q&R emulators, "Material > Menus" in demo app
Change-Id: I1fdef0c3e2b43d987f39e755e7a2cc2e00f838c5