Improve accessibility of Picker
Add Picker overload with onSelected and contentDescription parameters. onSelected parameter enables semantics to update the selected Picker in multi-Picker screens (otherwise, clicking the read-only Picker has no effect). contentDescription parameter sets the content for screen-readers for the currently selected option.

Bug: 232522757
Test: androidx.wear.compose.material.test
Relnote: "We have made accessibility improvements to Picker so that multi-picker screens are navigable with screen readers and the content description is accessible"
Change-Id: I64edb15386b8e92e4cc33b6a0bb816d333dba961