Fix cursor being unable to drag when VisualTransformation is applied to CoreTextField.
CoreTextField compares the old and new TextFieldValues and resets the selection or cursor if values do not match. However, this assumptions fails when there is a VisualTransformation because transformed text always becomes different compared to the original text. Furthermore, there was a forgotten step to convert drag position from transformed text to original text before creating a new TextFieldValue.

This CL adds originalText field to TextFieldState to enable the comparison to original text.

Test: TextFieldSelectionTest.kt
Fix: b/210192858
Fix: b/228340264
Change-Id: Iccc945d5714c654bf7c0d560adcf24474a1672a8