Added SurfaceControlCompat tests to test across all API levels for SurfaceControl.
Added tests for SurfaceControlCompat to ensure correctness among API
levels. This leads to small changes in the API to ensure similarity with
the SDK/NDK. SetBuffer releaseCallback now has a default value of null,
to mimic the SDK methods that allow not having to pass in a release
callback.

Relnote: "SurfaceControlCompat tests added to ensure correctness among
API levels."

Bug: b/234791586
Test: ./gradlew :graphics:graphics-core:connectedAndroidTest

Change-Id: I2f4cfa7b5102348c2e7f410f1061db06e86c0d7e