Lookahead API improvements
This CL introduces a new LookaheadScope that replaces LookaheadLayout and
allows the LookaheadScope to be created without needing an
additional Layout. This enables children of LookaheadScope to
be directly managed by the parent's measure/layout logic.

The new LookaheadScope also supports query of scope coordinates during
placement, as well as conversion from LayoutCoordinates to
LookaheadCoordinates.

The IntermediateLayoutModifier is now a LayoutModifierNode that has both
measure and intermediateMeasure implementation for the lookahead pass
and main pass respectively.

Modifier.intermediateLayout no longer requires a LookaheadScope, it will
instead implicitly create a LookaheadScope if it's not already in
a LookaheadScope.

New IntermediateMeasureScope that conveniently provides LookaheadScope,
CoroutineScope and MeasureScope to the measure logic in intermediateLayout.

Future work:

1) An API will be needed for devs to express whether the intermediate
 changes have finished. The implementation for wiring that up to snap to
 lookahead is already in place.

2) Ideally, the intermediateLayout modifier API should not require the
use of composed {}.

In an effort to reduce churns for developers, both 1) an 2) are deferred
till Modifier.Node Delegate design is finished.

Relnote: "
- Modifier.intermediateLayout now doesn't require an explicit LookaheadScope.
The measure block in intermediateLayout has IntermediateMeasureScope as
receiver, which provides convenient CoroutineScope, LookaheadScope and
MeasureScope.

- LookaheadLayout has been replaced by LookaheadScope, which is no longer
a Layout. This allows chid content in a LookaheadScope to be directly
controlled by parent's MeasurePolicy.
"

Test: Included

Change-Id: Ibe2e5e20a833d3824213e86e5c1e155f65647ad7