Introduce ExistingPeriodicWorkPolicy.Update
It is very similar to previously existing
ExistingPeriodicWorkPolicy.REPLACE.

Test: WorkManagerImplTest + integration test-app
Relnote: "`ExistingPeriodicWorkPolicy.UPDATE` was
added. This policy allows to update a periodic work
by the name. It is similar existing `REPLACE`, but
is less intrusive: it doesn't cancel a worker if it
is currently running, it preserves enqueue time, so
initial delay and period are calculated from original
enqueue time, rather than update time. `REPLACE` was
deprecated to decrease a confusion between very similarly
named `REPLACE`and `UPDATE`. If you still want to keep
previous semantics of `REPLACE`,
newly added `CANCEL_AND_REENQUEUE` identical to `REPLACE`
can be used."
bug: 219446409

Change-Id: I985edc2cf7ac7c316932185f3fd0a3ca6660f2dc