Make jump invalidation in PageFetcherSnapshot idempotent
Current jump handler operates on both append and prepend hint flows and checks for jumps on both flows with collectLatest.
This means it is possible for both prepend/append to each trigger multiple jumps on the same generation.

Jump invalidation should be triggered only once per generation. This means we need to merge the two flows and then terminate collection and invalidate on first jump hint.

Test: ./gradlew paging:paging-common:test
Bug: 235264742
Change-Id: I0b290477433c4e738e4f42060695eb3bf5d44e92