Update MediaRouteDescriptor using platform data
The previous attempt to do this was reverted in aosp/2127512.
That attempt was problematic because it entirely discarded the
group descriptor data saved under KEY_GROUP_ROUTE.

This version of the change retains the group descriptor data
saved under KEY_GROUP_ROUTE, overwriting only the fields for
which the platform should be the source of truth.

Bug: 236100132
Test: Manual Cast testing
Change-Id: I5112f4a1057583124abd89d2d73c6a1f52391bfc